C#: Re-factor the processing of all files into a helper class and make sure to only do one pass of the files.
